requirements.txt
amqp
apscheduler
celery
Django>=1.8,<1.9
djangorestframework
django-dirtyfields
django-fsm
django-haystack
django-reversion
django-transaction-hooks
drf-haystack
drf-nested-routers
kombu
Pillow
psycopg2
pytz
PyYAML
requests
SQLAlchemy
